refactor: `./internal` -> `./sdk/internal`
Since everything else is nested under `./sdk`, this may as well be too?
