+++ sdk/internal/test/base64.go
