+++ sdk/internal/test/metadata_stub.go
