+++ sdk/internal/test/msi_stub.go
